+++ light-reading-meter/Configs/Development.xcconfig
@@ -1,0 +1,10 @@
+//
+//  Config.xcconfig
+//  light-reading-meter
+//
+//  Created by Oscar Rivera Moreira on 15/2/24.
+//
+
+// Configuration settings file format documentation can be found at:
+// https://help.apple.com/xcode/#/dev745c5c974
+API_URL=localhost:3000
