+++ light-reading-meter/Features/Home/ViewModels/HomeViewModel.swift
@@ -12,21 +12,25 @@ class HomeViewModel: ObservableObject {
     @Published var myMeters: [Meter] = []
     @Published var showErrorAlert = false
     @Published var errorMessage: String?
+    @Published var isLoading: Bool = false
     
     init() {
+        self.isLoading = true
+        
         fetchMeters()
     }
     
     func fetchMeters() {
         MeterService.shared.getMeters { [weak self] success, meters, error in
-            if success {
-                DispatchQueue.main.async {
-                    self?.myMeters = meters
-                }
-            } else {
-                print("Error: \(error ?? "Unknown error")")
+            DispatchQueue.main.async {
+                sleep(1)
+                self?.isLoading = false
                 
-                DispatchQueue.main.async {
+                if success {
+                    self?.myMeters = meters
+                } else {
+                    print("Error: \(error ?? "Unknown error")")
+                    
                     self?.showErrorAlert = true
                     self?.errorMessage = NSLocalizedString("something_went_wrong", comment: "")
                 }
